# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TransJakarta Routes API
  description: |
    A RESTful API application built with Go (Gin framework) to manage TransJakarta BRT route information, stops/terminals, vehicles, and user reports.

    ## Features
    - User Authentication (Email/Password and OAuth)
    - Guest Access (read-only routes, stops, vehicles; create reports)
    - Role-Based Access Control (Common users and Admins)
    - Route Management
    - Stop/Terminal Management
    - Vehicle Management
    - Reporting System
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/v1/reports/{id}/comments:
    get:
      tags:
        - Comments
      summary: Get comments for a report
      description: Get all threaded comments for a specific report
      operationId: getReportComments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                success: true
                data:
                  comments:
                    - id: 1
                      report_id: 1
                      user_id: 2
                      content: "Great report!"
                      upvotes: 5
                      downvotes: 0
                      created_at: "2024-01-01T00:00:00Z"
                      replies: []
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags:
        - Comments
      summary: Add comment to report
      description: Add a comment to a report (can be a reply to another comment)
      operationId: createComment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/reports/{id}/react:
    post:
      tags:
        - Reactions
      summary: React to a report
      description: Add or toggle an upvote/downvote reaction on a report
      operationId: reactToReport
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactRequest"
      responses:
        "200":
          description: Reaction updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags:
        - Reactions
      summary: Remove reaction from report
      description: Remove your reaction from a report
      operationId: removeReportReaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reaction removed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/comments/{id}:
    put:
      tags:
        - Comments
      summary: Update comment
      description: Update your own comment
      operationId: updateComment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommentRequest"
      responses:
        "200":
          description: Comment updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags:
        - Comments
      summary: Delete comment
      description: Delete your own comment
      operationId: deleteComment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comment deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/comments/{id}/react:
    post:
      tags:
        - Reactions
      summary: React to a comment
      description: Add or toggle an upvote/downvote reaction on a comment
      operationId: reactToComment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactRequest"
      responses:
        "200":
          description: Reaction updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags:
        - Reactions
      summary: Remove reaction from comment
      description: Remove your reaction from a comment
      operationId: removeCommentReaction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reaction removed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/leaderboard:
    get:
      tags:
        - Leaderboard
      summary: Get leaderboard
      description: Get top users by reputation points
      operationId: getLeaderboard
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 100
      responses:
        "200":
          description: Leaderboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                success: true
                data:
                  leaderboard:
                    - id: 1
                      username: "user1"
                      reputation_points: 500
                      level: "expert"
                  limit: 10
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/users/{id}/profile:
    get:
      tags:
        - Leaderboard
      summary: Get user profile
      description: Get user profile with reputation points, level, and badges
      operationId: getUserProfile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                success: true
                data:
                  user:
                    id: 1
                    username: "user1"
                    email: "user1@example.com"
                  reputation_points: 250
                  level: "trusted"
                  badges:
                    - id: 1
                      badge_id: 1
                      earned_at: "2024-01-01T00:00:00Z"
                      badge:
                        id: 1
                        name: "First Report"
                        description: "Submitted your first report"
                        icon: "🎯"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/badges:
    get:
      tags:
        - Badges
      summary: Get all badges
      description: Get list of all available badges
      operationId: getAllBadges
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of badges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                success: true
                data:
                  badges:
                    - id: 1
                      name: "First Report"
                      description: "Submitted your first report"
                      icon: "🎯"
                      criteria_type: "reports_accepted"
                      criteria_value: 1
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    CreateCommentRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          example: "This is a helpful comment"
        parent_id:
          type: integer
          nullable: true
          description: ID of parent comment if this is a reply
          example: null
    UpdateCommentRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          example: "Updated comment content"
...
    ReactRequest:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum: [upvote, downvote]
          example: upvote
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message here
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error: "Invalid request parameters"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error: "Unauthorized access"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error: "Access forbidden"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error: "Resource not found"